server/requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
pydantic==2.12.5
google-genai
python-dotenv==1.2.1
python-multipart==0.0.21
requests==2.32.5
huggingface_hub